+++ packages/stream_chat_flutter/lib/src/extension.dart
@@ -225,3 +225,12 @@ extension UserListX on List<User> {
     return entries.map((e) => e.key).toList(growable: false);
   }
 }
+
+/// Extensions on [Uri]
+extension UriX on Uri {
+  /// Return the URI adding the http scheme if it is missing
+  Uri get withScheme {
+    if (hasScheme) return this;
+    return Uri.parse('http://${toString()}');
+  }
+}