+++ packages/stream_chat_persistence/lib/src/stream_chat_persistence_client.dart
@@ -1,5 +1,4 @@
 import 'package:flutter/foundation.dart';
-import 'package:logging/logging.dart' show LogRecord;
 import 'package:mutex/mutex.dart';
 import 'package:stream_chat/stream_chat.dart';