+++ packages/stream_chat_flutter/CHANGELOG.md
@@ -5,6 +5,7 @@
 - [[#1067]](https://github.com/GetStream/stream-chat-flutter/issues/1067): Fix name text overflow in reaction card.
 - [[#842]](https://github.com/GetStream/stream-chat-flutter/issues/842): show date divider for first message.
 - Loosen up url check for attachment download.
+- Use `ogScrapeUrl` for LinkAttachments.
 
 ## 3.6.1